mandatory fixes
